Fix constant propagation pass
This pass was broken in a number of ways, as we were not generating
asserts whenever we took it, even though we need to.

While doing so, we found that the analysis we were using for choosing
whether to generate asserts or not for dynamic shapes was completely
broken. We implement a couple heuristics that cover quite a few common
cases in the pass, and we always generate upper bounds for the indirect
indexing case.

In the future, we should be able to still vectorise code that goes
through the fallback if:
- We track which variables do not depend on data
- We write a vectorised `TORCH_CHECK`.

ghstack-source-id: 1fa0dcfdcc431e457f92b3bd81d53083cab93d70
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114471