Patch ARM Half use_gemv_fast_path gate to avoid kernel duplication
Summary: The existing code didn't gate the fast path, so the fast path had to duplicate the stock kernel. Now we gate it and delete the duplicate kernel.

Test Plan: Existing tests. Flipped the TORCH_INTERNAL_ASSERT_DEBUG_ONLY to non-debug and forced to fail (locally) to make sure we had test coverage.

ghstack-source-id: 550e23e7fecd688d79722126ea80cb74604d7b9e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/127478