[reland] Refresh OpOverloadPacket if a new OpOverload gets added
If a user accesses an OpOverloadPacket, then creates a new OpOverload,
then uses the OpOverloadPacket, the new OpOverload never gets hit. This
is because OpOverloadPacket caches OpOverloads when it is constructed.

This PR fixes the problem by "refreshing" the OpOverloadPacket if a new
OpOverload gets constructed and the OpOverloadPacket exists.

Test Plan:
- new tests

This is the third land attempt. The first one was reverted for breaking
internal tests, the second was reverted for being erroneously suspected
of causing a perf regression.

ghstack-source-id: 5e42a8e75c2cf84a6bc229f0b6e7647710f7305f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/128000