Improve unbacked reasoning involving has internal overlap
Fixes https://github.com/pytorch/pytorch/issues/122477
Partially addresses https://github.com/pytorch/pytorch/issues/116336

This PR is slightly overkill: not only does it disable the overlap test
when there are unbacked SymInts, it also improves the is non-overlapping
and dense test for some more unbacked situations.  We technically don't
need the latter change, but I was already deep in the sauce and just
went ahead and did it.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: 155bdc94323677fa1066ee343e11d93b2d17d2da
Pull Request resolved: https://github.com/pytorch/pytorch/pull/128332