[inductor] fx graph cache: memoize devices to make cache key calculation more predictable
Summary: I've seen this issue once in the wild and oulgen was able to repro in a unit test. The problem is this:
- We're using pickle to turn everything related to the FX graph cache key into a byte stream, then hashing the bytes to compute the cache key.
- Pickle is optimized to avoid serializing the same ID more than once; it instead drops a reference to a previously-pickled object if it encounters the same ID.
- That pickle behavior means that we can see different cache keys if an object id appears more than once in the hashed objects vs. being functionally equivalent but distinct objects.

The cases I've investigated only involve the torch.device objects in the tensor graph args. That is, we may compile a graph with two tensor each referencing torch.device('cpu'). In one run, those devices may reference the same object; in another they may reference distinct but equivalent objects. In practice, my observation is that the compiler is largely deterministic and this situation is rare. I've seen cache misses on a real benchmark only when enabling/disabling FakeTensor caching in order to introduce different execution paths that otherwise produce the same fx graph. But the failing unit test seems to be enough motivation to put in some remediation.

I don't really love this solution, but I've failed to find another way to make the pickling phase robust to these kinds of changes, e.g., by changing the protocol version or by overriding internal methods (which would also be gross). But I'm definitely open to other creative ideas.

ghstack-source-id: e97b7fc52b22262b7f23a2b2b67df8366665aabd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/128366