Remove some implications from the static_eval pattern matcher
Horace reported a outrageous increase in compilation time in
`compile(gpt-fast, dynamic=True)`

The first PR in this stack brings down that time from 42 min to 6 min.

This PR it down from 6.10 min to 4.12 min. For reference, if we comment
out `get_implications`, we get a time of 3.44 min. 30s difference is not
nothing, but it should be good enough for now.

We should be able to remove this as, with the new canonicalisation, we
have that `a < b` and `-a > -b` should be canonicalised to the same
expression (if SymPy does not interfere too much).

ghstack-source-id: 0bf26eb66059e6dda354383b8090927c5b927509
Pull Request resolved: https://github.com/pytorch/pytorch/pull/128500